+++ src/confusius/registration/__init__.py
@@ -4,6 +4,7 @@ from confusius.registration._progress import RegistrationProgressPlotter
 from confusius.registration.affines import (
     compose_affine,
     decompose_affine,
+    get_euler_xyz_from_rotation_matrix,
 )
 from confusius.registration.bspline import (
     invert_displacement_field,
@@ -30,6 +31,7 @@ __all__ = [
     "sample_displacement_field_like",
     "compose_affine",
     "decompose_affine",
+    "get_euler_xyz_from_rotation_matrix",
     "invert_displacement_field",
     "register_volume",
     "resample_volume",